backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
langchain==0.1.4
langchain-community==0.0.16
langchain-google-genai==0.0.6
sentence-transformers==2.3.1
pdfplumber==0.10.4
python-docx==1.1.0
python-pptx==0.6.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.3
